NetKVM: Virtqueue: Don't touch hardware upon suprise removal
Currently when we recieve a suprise remove IRP from Windows we touc
the hardware of the device which is an invalid memory. This doesn't
cause BSOD in HCK/HLK tests as the device is still in qemu and memory is
still valid.

However, when performing hibernate on a given client and
then - while the client is off - we remove the virtio-net from the
command line and boot the client, a BSOD occurs as a result of this
illegal access to memory.

This commit solves this issue by registering the paths for notifications from
the state machine and preventing the driver from touching hardware upon a
suprise removal notification from the state machine.

Signed-off-by: Sameeh Jubran <sameeh@daynix.com>