NetKVM: CX: Extend SendControlMessage bufs > 2
The Control Queue can handle as much buffers as we want to, currently we
only support 2 buffers from the driver and both of them are from the
driver to the device.

According to section "5.1.6.5 Control Virtqueue" of the virtio spec, the
send convention to the CX path is:

struct virtio_net_ctrl {
        u8 class;
        u8 command;
        u8 command-specific-data[];
        u8 ack;
};

How it is implemented:

* Send a scatter gather to the device which includes:
 ** virtio_net_ctrl_hdr as an output (class and command fields)
 ** 0 or more output buffers
 ** 0 or more input buffers
 ** virtio_net_ctrl_ack as an input buffer

The actual difference between input and output buffer is that output
buffer is writeable by the device.

This commit extends this number from 2 + 2 to 8 + 2. Note that the 2
buffers are the structs that we always send with each control command -
virtio_net_ctrl_hdr and virtio_net_ctrl_ack).

Moreover this commit enables us to send buffers as an output buffer, a
buffer which we send for the device and expect to receive back on
completion.

Signed-off-by: Sameeh Jubran <sameeh@daynix.com>